chore(renovate): update code
* Migrate Web framework (Spark) to Javalin
* Migrate DI (Kodein) to Koin
* Migrate DB (Mongo/Jongo) to vanilla in memory storage
* Fix build
* Fix tests

+++ src/main/kotlin/org/github/mbarberot/mtg/grimoire/business/searches/SearchMetadata.kt
@@ -1,8 +1,8 @@
 package org.github.mbarberot.mtg.grimoire.business.searches
 
 data class SearchMetadata(
-        val resultsCount: Long,
+        val resultsCount: Int,
         val pageSize: Int,
-        val currentPage: Long,
+        val currentPage: Int,
         val query: String
 )
